requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
logging==0.4.9.6
asyncio==3.4.3
pytz==2024.1
aiohttp==3.9.3
cryptography==42.0.5
